requirements.txt
# extractor
requests
pycryptodome

# extract
prettytable
click

# web
flask
flask-cors
python-dotenv
